requirement.txt
django
django-braces
requests